requirements.txt
aiohttp-3.6.2
async-timeout-3.0.1
attr-0.3.1
attrs-19.3.0
chardet-3.0.4
idna-2.9
multidict-4.7.6
pymongo-3.10.1
urllib3-1.25.9
yarl-1.4.2